requirements.txt
acres==0.5.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.1
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
cssselect2==0.8.0
cycler==0.12.1
etelemetry==0.3.1
filelock==3.20.0
fonttools==4.61.0
fsspec==2025.12.0
hf-xet==1.2.0
httplib2==0.31.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.6
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pillow==12.0.0
prov==2.1.1
puremagic==1.30
pycparser==2.23
pydot==4.0.1
pydyf==0.12.1
PyMuPDF==1.26.6
pyparsing==3.2.5
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.4
PyYAML==6.0.3
rdflib==7.5.0
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scipy==1.16.3
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
sympy==1.14.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
traits==7.0.2
transformers==4.57.3
triton==3.5.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==67.0
webencodings==0.5.1
wheel==0.45.1
zopfli==0.4.0
